@@ -22,27 +22,27 @@
     "webpack"
   ],
   "dependencies": {
-    "antd": "0.9.0",
+    "antd": "~0.10.1",
     "d3": "^3.5.5",
     "events": "^1.0.2",
     "flux": "^2.0.1",
-    "gregorian-calendar": "3.x",
-    "gregorian-calendar-format": "3.x",
     "jquery": "^1.11.3",
     "jsonp": "^0.2.0",
     "object-assign": "^2.0.0",
     "querystring": "^0.2.0",
-    "rc-calendar": "~3.12.4",
-    "rc-dialog": "~4.0.7",
-    "react": "~0.13.0",
-    "react-art": "^0.13.0",
-    "react-intl": "~1.2.0",
-    "react-router": "1.0.0-rc1"
+    "rc-animate": "~2.0.2",
+    "rc-dialog": "~5.3.0",
+    "react": "~0.14.3",
+    "react-art": "~0.14.0",
+    "react-dom": "~0.14.3",
+    "react-router": "1.0.x"
   },
   "devDependencies": {
+    "history": "1.x",
+    "react-addons-test-utils": "~0.14.3",
     "babel-core": "~5.8.24",
     "expect.js": "~0.3.1",
-    "precommit-hook-eslint": "~3.0.0",
+    "pre-commit": "1.x",
     "rc-server": "3.x",
     "rc-tools": "4.x",
     "replace": "~0.3.0",
@@ -52,14 +52,13 @@
     "build": "rc-tools run build",
     "gh-pages": "rc-tools run gh-pages",
     "pre-gh-pages": "sh ./scripts/pre-gh-pages.sh",
-    "start": "node --harmony node_modules/.bin/rc-server",
+    "start": "rc-server",
     "pub": "rc-tools run pub",
     "lint": "rc-tools run lint",
     "karma": "rc-tools run karma",
-    "saucelabs": "node --harmony node_modules/.bin/rc-tools run saucelabs",
-    "browser-test": "node --harmony node_modules/.bin/rc-tools run browser-test",
-    "browser-test-cover": "node --harmony node_modules/.bin/rc-tools run browser-test-cover",
-    "validate": "npm ls"
+    "saucelabs": "rc-tools run saucelabs",
+    "browser-test": "rc-tools run browser-test",
+    "browser-test-cover": "rc-tools run browser-test-cover"
   },
   "pre-commit": [
     "lint"
