@@ -1,4 +1,4 @@
-const Component = require('./Component');
-const React = require('react');
+import React from 'react';
+import Component from './Component';
 const appData = window.appData;
 React.render(<Component {...appData}/>, document.body);
