@@ -1,6 +1,6 @@
-const React = require('react');
-const Component = require('./Component');
-
+import React from 'react';
+import Component from './Component';
+import ReactDOM from 'react-dom';
 const Test = React.createClass({
   render() {
     const props = this.props;
@@ -11,14 +11,9 @@ const Test = React.createClass({
   },
 });
 
-const test = React.render(<Test id1="id1" id2="id2" key1="key1" key2="key2"/>, document.getElementById('__react-content'));
+ReactDOM.render(<Test id1="id1" id2="id2" key1="key1" key2="key2"/>, document.getElementById('__react-content'));
 
 setTimeout(() => {
   console.log('*******************************');
-  test.setProps({
-    id1: 'id11',
-    id2: 'id2',
-    key1: 'key11',
-    key2: 'key2',
-  });
+  ReactDOM.render(<Test id1="id1" id2="id2" key1="key11" key2="key2"/>, document.getElementById('__react-content'));
 }, 1000);
