@@ -1,15 +1,15 @@
-
 let container;
-require('../../assets/load-on-demand/async.css');
-const Dialog = require('rc-dialog');
-require('rc-dialog/assets/bootstrap.css');
-const React = require('react');
+import '../../assets/load-on-demand/async.css';
+import 'rc-dialog/assets/index.css';
+import Dialog from 'rc-dialog';
+import React from 'react';
+import ReactDOM from 'react-dom';
 exports.show = () => {
   if (!container) {
     container = document.createElement('div');
     document.getElementById('__react-content').appendChild(container);
   }
-  React.render(<Dialog style={{width: 400}}>
+  ReactDOM.render(<Dialog style={{width: 400}} title="test dialog" visible closable={false}>
     <div className="my-content">loaded</div>
-  </Dialog>, container).show();
+  </Dialog>, container);
 };
